+++ .storybook/main.js
@@ -1,6 +1,6 @@
 module.exports = {
   stories: [
-    '../src/*/stories.js',
+    '../src/*/**/stories.js',
   ],
   addons: [
     '@storybook/addon-docs',