+++ src/Carousel/index.js
@@ -1,5 +1,4 @@
 import React, { useState } from 'react'
-import Icon from '../Icon'
 import ButtonIcon from '../ButtonIcon'
 import cx from 'classnames'
 import {default as PT} from 'prop-types'
@@ -26,9 +25,8 @@ const Carousel = ({ children, className }) => {
         <ButtonIcon
           onClick={() => moveNext(-1)}
           disabled={!canMoveLeft}
-        >
-          <Icon name="chevronLeft" size="big" />
-        </ButtonIcon>
+          icon="chevronLeft"
+        />
         {children.map((item, index) => {
           return index === itemSelected ? React.cloneElement(item, {
             key: index
@@ -37,9 +35,8 @@ const Carousel = ({ children, className }) => {
         <ButtonIcon
           onClick={() => moveNext()}
           disabled={!canMoveRight}
-        >
-          <Icon name="chevronRight" size="big" />
-        </ButtonIcon>
+          icon="chevronRight"
+        />
       </div>
       <div className="ola_carousel-dots">
         {children.map((item, i) => (