+++ src/Field/index.js
@@ -1,9 +1,28 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { default as PT } from 'prop-types'
 import cx from 'classnames'
 
-const Field = ({ id, label, hint, error, description, disabled, className, children }) => {
+const Field = ({ id, label, hint, counter, error, description, disabled, className, children }) => {
   let elements = React.Children.toArray(children)
+  const childProps = elements[0].props || {}
+  const [localValue, setLocalValue] = useState(childProps.value || '')
+
+  const extendChildProps = {
+    id,
+    error,
+    disabled,
+  }
+
+  if (counter && childProps.maxLength) {
+    extendChildProps.onInput = (e) => {
+      setLocalValue(e.target.value)
+
+      if (childProps.onInput && typeof childProps.onInput === 'function') {
+        childProps.onInput(e)
+      }
+    }
+  }
+
   return (
     <div className={cx('ola_field', { 'is-invalid': error }, { 'is-disabled': disabled }, className)}>
       <label htmlFor={id} className="ola_field-label">
@@ -11,10 +30,15 @@ const Field = ({ id, label, hint, error, description, disabled, className, child
         {hint && <span className="ola_field-hint">{hint}</span>}
       </label>
       <div className="ola_field-input">
-        {React.cloneElement(elements.shift(), { id: id, error, disabled })}
+        {React.cloneElement(elements.shift(), extendChildProps)}
       </div>
       {elements}
-      {description && <p className={cx({ 'ola_field-error': error, 'ola_field-description': !error })}>{description}</p>}
+      {description && <p className={cx({ 'ola_field-error': error, 'ola_field-description': !error })}>
+        {description}
+        {
+          counter && childProps.maxLength && <span className="ola_field-counter">{localValue.length}/{childProps.maxLength}</span>
+        }
+      </p>}
     </div>
   )
 }
@@ -23,7 +47,8 @@ Field.defaultProps = {
   error: false,
   hint: null,
   description: null,
-  disabled: false
+  disabled: false,
+  counter: false,
 }
 
 Field.propTypes = {
@@ -39,6 +64,8 @@ Field.propTypes = {
     PT.node,
     PT.string
   ]),
+  /** Show a char counter visually (needs a children like Input with a maxLength prop) */
+  counter: PT.bool,
   /** Description */
   description: PT.oneOfType([
     PT.arrayOf(PT.node),