+++ src/Input/index.js
@@ -1,28 +1,43 @@
 import React from 'react'
 import { default as PT } from 'prop-types'
 import cx from 'classnames'
+import ButtonIcon from '../ButtonIcon'
 
-const Input = React.forwardRef(({ className, error, type, ...props }, ref) => {
-  return type === 'textarea' ? (
-    <textarea
-      ref={ref}
-      type={type}
-      className={cx('ola_input', { 'is-invalid': error }, className)}
-      {...props}
-    />
-  ) : (
-    <input
-      type={type}
-      ref={ref}
-      className={cx('ola_input', { 'is-invalid': error }, className)}
-      {...props}
-    />
-  )
+const Input = React.forwardRef(({ className, error, type, icon, ...props }, ref) => {
+  const Element = type === 'textarea' ? 'textarea' : 'input'
+
+  if (type === 'textarea' && !props.rows) {
+    props.rows = 3
+  }
+
+  if (type !== 'textarea') {
+    props.type = type
+  }
+
+  const createElement = () => <Element
+    ref={ref}
+    className={cx('ola_input', { 'is-invalid': error, 'is-icon': !!icon }, className)}
+    {...props}
+  />
+
+  if (icon) {
+    return (
+      <div className="ola_input-wrapper">
+        {createElement()}
+        <div className="ola_input-icon">
+          {icon}
+        </div>
+      </div>
+    )
+  }
+ 
+  return createElement()
 })
 
 Input.defaultProps = {
   type: 'text',
   className: null,
+  icon: null,
   error: false,
 }
 
@@ -47,6 +62,11 @@ Input.propTypes = {
   className: PT.string,
   /** Input is invalid */
   error: PT.bool,
+  /** Optional ButtonIcon to show on the top right corner. */
+  icon: PT.oneOf([
+    ButtonIcon
+  ]),
+  rows: PT.number,
 }
 
 Input.displayName = 'Input'