+++ src/Keyword/test.js
@@ -1,7 +1,6 @@
 import React from 'react'
 import Keyword from './'
 import ButtonIcon from '../ButtonIcon'
-import Icon from '../Icon'
 import renderer from 'react-test-renderer'
 
 it('Default Keyword', () => {
@@ -15,6 +14,6 @@ it('Keyword with extraClass', () => {
 })
 
 it('Keyword with button', () => {
-  const tree = renderer.create(<Keyword className="test" button={<ButtonIcon><Icon name="close" size="small" /></ButtonIcon>}>Test</Keyword>).toJSON()
+  const tree = renderer.create(<Keyword className="test" button={<ButtonIcon icon="close" />}>Test</Keyword>).toJSON()
   expect(tree).toMatchSnapshot()
 })