+++ src/Segment/test.js
@@ -1,27 +1,14 @@
 import React from 'react'
-import Button from '../Button'
 import Segment from './'
 import renderer from 'react-test-renderer'
 
-it('Segment with two buttons', () => {
+it('Default Segment', () => {
   const tree = renderer
     .create(
       <Segment>
-        <Button variant="secondary">Main button</Button>
-        <Button variant="secondary">Secondary button</Button>
-      </Segment>
-    )
-    .toJSON()
-
-  expect(tree).toMatchSnapshot()
-})
-
-it('Segment with two buttons and an extra class and one button pressed', () => {
-  const tree = renderer
-    .create(
-      <Segment className="extra-class">
-        <Button variant="secondary" selected>Main button</Button>
-        <Button variant="secondary">Secondary button</Button>
+        <Segment.Button selected>One &lt;Segment.Button&gt;</Segment.Button>
+        <Segment.Button>I am a button</Segment.Button>
+        <Segment.Button as="a" href="https://www.marketgoo.com/" target="_blank">I am an anchor</Segment.Button>
       </Segment>
     )
     .toJSON()