+++ src/Select/stories.js
@@ -6,12 +6,14 @@ export default {
   component: Select,
   args: {
     options: [
-      { value: '', label: 'I’m not sure' },
+      { value: '', label: 'I\'m not sure' },
       { value: 'chocolate', label: 'Chocolate' },
       { value: 'strawberry', label: 'Strawberry' },
       { value: 'vanilla', label: 'Vanilla' },
       { value: '', label: 'None' }
-    ]
+    ],
+    disabled: false,
+    error: false
   },
   argTypes: {
     onClick: { action: 'clicked' },
@@ -20,3 +22,13 @@ export default {
 }
 
 export const Base = (args) => <Select {...args}/>
+
+export const Error = (args) => <Select {...args}/>
+Error.args = {
+  error: true
+}
+
+export const Disabled = (args) => <Select {...args}/>
+Disabled.args = {
+  disabled: true
+}