+++ src/Tooltip/stories.js
@@ -2,7 +2,6 @@ import React from 'react'
 
 import Tooltip from './'
 import Button from '../Button'
-import Icon from '../Icon'
 import ButtonIcon from '../ButtonIcon'
 
 export default {
@@ -27,9 +26,7 @@ export const WithButtonIconHover = (args) => (
     {...args}
     hover
     trigger={
-      <ButtonIcon as="span">
-        <Icon name="help" />
-      </ButtonIcon>
+      <ButtonIcon as="span" icon="help" />
     }
   >
     Content
@@ -40,9 +37,7 @@ export const WithButtonIcon = (args) => (
   <Tooltip
     {...args}
     trigger={
-      <ButtonIcon as="span">
-        <Icon name="help" />
-      </ButtonIcon>
+      <ButtonIcon as="span" icon="help" />
     }
   >
     Content
@@ -75,9 +70,7 @@ export const InAScrollableElement = () => (
     <div style={{ padding: '300px' }}>
       <Tooltip
         trigger={
-          <ButtonIcon as="span">
-            <Icon name="help" />
-          </ButtonIcon>
+          <ButtonIcon as="span" icon="help" />
         }
       >
         Lorem ipsum